@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
     <DebugType>full</DebugType>
@@ -9,7 +9,7 @@
   </PropertyGroup>
 
   <PropertyGroup>
-    <PackageId>Fibula.Scheduling.Contracts</PackageId>
+    <PackageId>Fibula.EventsEngine.Contracts</PackageId>
     <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
     <Copyright>2020</Copyright>
@@ -17,16 +17,12 @@
     <Company>The Fibula Project</Company>
     <Product>Fibula MMO</Product>
     <PackageProjectUrl>https://fibula-mmo.github.io/</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/Fibula-MMO/fibula-scheduling</RepositoryUrl>
+    <RepositoryUrl>https://github.com/Fibula-MMO/fibula-events-engine</RepositoryUrl>
     <RepositoryType>GitHub</RepositoryType>
-    <Version>1.1.3</Version>
+    <Version>1.0.0</Version>
     <NeutralLanguage>en</NeutralLanguage>
-    <Description>Contains Abstractions for the scheduling framework used in the Fibula project.</Description>
+    <Description>Contains abstractions for the events engine framework used in the Fibula project.</Description>
   </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Remove="Abstractions\IEventCondition.cs" />
-  </ItemGroup>
 
   <ItemGroup>
     <None Remove="stylecop.json" />
@@ -42,7 +38,7 @@
 
   <ItemGroup>
     <PackageReference Include="Fibula.Utilities.Validation" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.0" />
     <PackageReference Include="StyleCop.Analyzers" Version="1.1.118">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
